requirements.txt
random
kafka
pyspark.sql.functions
pyspark.sql.types
pyspark.sql
mysql.connector
pandas
time
random